Refactor Schedule Write DAOs
